Add files via upload
This is made with Open Office. What is important if peoples modify this file is that it is still made with Open Office and no other softwares. What i would like is that peoples that know Solarus Quest Maker, ask to them self a simple question: "How to do this/that?". Brief put them self in the shoes of a true beginner, that dont know how work Solarus Quest Maker, how to program in LUA; and indicate each steps with details.

The document file, the properties of it can not be modified. This mean the descriptions, the size of the document, the background image, the different zones. So please dont modify it. Just add some text and image into the pages.

The only thing that i would like, is that you would do some printscreens of what you do and put some text with many details. The image quality would be of 96dpi at leass and not much. Like 96dpi or 120dpi or 300dpi max. Because when with Open Office the export to pdf icon is clicked, if the pdf file is more then 5MB; then it is difficult